# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Administrateur créé avec succès"
          },
          "400": {
            "description": "Données invalides"
          },
          "401": {
            "description": "Non autorisé"
          },
          "403": {
            "description": "Accès interdit"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un nouvel administrateur (super admin uniquement)",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste des administrateurs récupérée avec succès"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer la liste des administrateurs",
        "tags": [
          "users"
        ]
      }
    },
    "/api/admin/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profil récupéré avec succès"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer son propre profil",
        "tags": [
          "users"
        ]
      }
    },
    "/api/admin/users/profile/password": {
      "put": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mot de passe changé avec succès"
          },
          "400": {
            "description": "Mot de passe actuel incorrect"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changer son propre mot de passe",
        "tags": [
          "users"
        ]
      }
    },
    "/api/admin/users/profile/avatar": {
      "put": {
        "operationId": "UsersController_updateAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar mis à jour avec succès"
          },
          "400": {
            "description": "Fichier manquant ou invalide"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour son propre avatar",
        "tags": [
          "users"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'administrateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Administrateur récupéré avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "404": {
            "description": "Administrateur non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un administrateur par son ID",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'administrateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Administrateur mis à jour avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "403": {
            "description": "Accès interdit"
          },
          "404": {
            "description": "Administrateur non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un administrateur (super admin uniquement)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'administrateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Administrateur supprimé avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "403": {
            "description": "Accès interdit"
          },
          "404": {
            "description": "Administrateur non trouvé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un administrateur (super admin uniquement)",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Aloha Secourisme",
    "description": "\n      API REST complète pour la gestion de l'application Aloha Secourisme.\n      \n      Cette API permet de gérer les formations, les inscriptions, les documents, \n      la galerie d'images et l'administration du système.\n      \n      ## Authentification\n      \n      La plupart des endpoints nécessitent une authentification par token JWT.\n      Pour vous authentifier, utilisez l'endpoint `/api/auth/login` pour obtenir un token,\n      puis incluez ce token dans l'en-tête d'autorisation de vos requêtes.\n      \n      ## Rôles\n      \n      - **Accès public** : Accessible sans authentification\n      - **Admin** : Nécessite d'être connecté avec un compte administrateur\n      - **Super Admin** : Nécessite d'être connecté avec un compte super administrateur\n      \n      ## Modules principaux\n      \n      - **Formations** : Gestion des sessions de formation\n      - **Inscriptions** : Gestion des inscriptions aux formations\n      - **Documents** : Gestion des documents téléchargeables\n      - **Galerie** : Gestion des images de la galerie\n      - **News** : Gestion des actualités\n      - **Admin/Users** : Gestion des utilisateurs administrateurs\n      - **Dashboard** : Statistiques et tableaux de bord\n      - **Settings** : Paramètres du système\n      - **Database** : Export/import de la base de données\n    ",
    "version": "1.0.0",
    "contact": {
...
    },
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Entrez votre token JWT ici"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com",
            "description": "Email de l'administrateur"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "description": "Mot de passe (min 8 caractères)"
          },
          "nom": {
            "type": "string",
            "example": "Dupont",
            "description": "Nom de famille"
          },
          "prenom": {
            "type": "string",
            "example": "Jean",
            "description": "Prénom"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "SUPER_ADMIN"
            ],
            "description": "Rôle de l'administrateur",
            "example": "ADMIN"
          }
        },
        "required": [
          "email",
          "password",
          "nom",
          "prenom",
          "role"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "CurrentPassword123!",
            "description": "Mot de passe actuel"
          },
          "newPassword": {
            "type": "string",
            "example": "NewPassword123!",
            "description": "Nouveau mot de passe (min 8 caractères)"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com",
            "description": "Email de l'administrateur"
          },
          "nom": {
            "type": "string",
            "example": "Dupont",
            "description": "Nom de famille"
          },
          "prenom": {
            "type": "string",
            "example": "Jean",
            "description": "Prénom"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "SUPER_ADMIN"
            ],
            "description": "Rôle de l'administrateur",
            "example": "ADMIN"
          },
          "password": {
            "type": "string",
            "example": "NewPassword123!",
            "description": "Nouveau mot de passe (min 8 caractères)"
          }
        }
      },
...
    }
  }
}